requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
oauthlib==3.3.1
passlib==1.7.4
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.35.0
setuptools==78.1.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
